refactor: organize business modules
