+++ src/app/favorites/components/favorite-card/favorite-card.component.html
