+++ src/app/favorites/components/favorite-card/favorites-card.component.spec.ts
