+++ src/app/favorites/favorites.component.html
