+++ src/app/favorites/favorites.component.spec.ts
