+++ src/app/favorites/favorites.component.ts
