+++ src/app/favorites/favorites.module.ts
@@ -4,7 +4,7 @@ import { HttpClient } from '@angular/common/http';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { httpLoaderFactory } from 'src/app/utils/http-loader.factory';
 import { FavoritesRoutingModule } from './favorites-routing.module';
-import { SharedModule } from '../../shared/shared.module';
+import { SharedModule } from '../shared/shared.module';
 import { AngularSvgIconModule } from 'angular-svg-icon';
 import { ReactiveFormsModule, FormBuilder } from '@angular/forms';
 import { FavoritesComponent } from './favorites.component';
