+++ src/app/favorites/services/favorite.facade.ts
