+++ src/app/items/components/item-card/item-card.component.spec.ts
