+++ src/app/items/services/items.facade.spec.ts
@@ -2,7 +2,7 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
 import { of } from 'rxjs';
 import { ItemsFacade } from './items.facade';
-import { IQueryParams } from '@api';
+import { IQueryParams } from 'src/app/items/api';
 import { ItemsService } from './items/items.service';
 import { itemMockModel } from '../mocks/item-mock.model';
 import { ItemModel } from '../models/item.model';
@@ -10,11 +10,11 @@ import { ItemsState } from '../state/items.reducer';
 import { Store, StoreModule } from '@ngrx/store';
 import * as fromItems from '../state/items.reducer';
 import { getItems, resetStateItems, setFavPropItems } from '../state/items.actions';
-import { AppFacade } from '../../../core/services/app.facade';
+import { AppFacade } from '../../core/services/app.facade';
 import { FormBuilder } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
-import { productMockModel } from '../../../core/mocks/product-mock.model';
-import { ProductModel } from '../../../core/models/product.model';
+import { productMockModel } from '../../core/mocks/product-mock.model';
+import { ProductModel } from '../../core/models/product.model';
 
 const itemRes: ItemModel[] = [
   itemMockModel,
